+++ FlexLib/Classes/ViewExt/UIView+Flex.m
@@ -75,6 +75,33 @@ FLEXSET(bgColor)
     }
 }
 
+FLEXSET(font)
+{
+    SEL sel = @selector(setFont:);
+    NSMethodSignature* sig = [self.class instanceMethodSignatureForSelector:sel];
+    if(sig == nil)
+    {
+        NSLog(@"Flexbox: no setFont: in class %@",[self class]);
+        return ;
+    }
+    
+    UIFont* font = fontFromString(sValue);
+    if(font==nil){
+        return;
+    }
+    
+    @try{
+        
+        NSInvocation* inv = [NSInvocation invocationWithMethodSignature:sig] ;
+        [inv setTarget:self];
+        [inv setSelector:sel];
+        [inv setArgument:&font atIndex:2];
+        
+        [inv invoke];
+    }@catch(NSException* e){
+        NSLog(@"Flexbox: setFont: called failed.");
+    }
+}
 FLEXSET(borderWidth)
 {
     CGFloat f = [sValue floatValue] ;